Internal, but disabled, TLS support for i386
This commit shows how internal TLS address lookup support could
be implemented for the i386 target.

Unfortunately, it doesn't work due to I386_GSBASE_REGNUM being
unavailable for Linux targets.  I looked at trying to access the
gsbase register via PTRACE_GET_THREAD_AREA, but did not understand
it well enough to finish it.  Since the i386 target is much less
important than it used to be, I gave up working on it.

I don't want to leave this disabled code in our sources, so I
will delete it in the next commit, however, this commit will be
in our git repo, so it'll be available for someone with sufficient
interest in the i386 target to look at.

Tested-By: Luis Machado <luis.machado@arm.com>
Approved-By: Luis Machado <luis.machado@arm.com>